requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-eventhub
azure-functions==1.7.2   
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30       
charset-normalizer==2.0.6
colorama==0.4.4
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
requests==2.26.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0